Add solutions for take home assignment 9
This commit introduces the solutions for the assignment 9 of the Data Structures and Algorithms course.

The first file is the solution for question 2 (x_22001158_q2.c), which implements a tree and a function to determine whether it is symmetric. The implementation uses a recursive approach to check the symmetry of a binary tree, while taking into account the root of the tree.

The second new file is the solution for question 1 (x_22001158_q1.c). This code calculates and displays the number of subordinates for each node in a tree. The node here denotes employees in a hierarchical organization. The code displays the total count of subordinates for each employee.