Create new DSA and SE assignment files
Added a new C program file (DSA/take_home_assignment_12/22001158_22001298.c) for DSA assignment. The program provides examples of various sorting algorithms including bubble and insertion sort, a function to identify the largest integer in an array, and a function to calculate the length of the longest increasing subsequence. Also, added a new use case diagram (SE/Assignment1/usecasediagram.svg) as part of the SE assignment.